Sharing Of Appliances
Currently the load per appliance is calculated for each person and just added together. 
This means that for example for the fridge the load is added multiple times. The load for the fridge will therefore be overestimated. 
To allow for sharing: the load calculation in 'residential.py' is adapted for each type of appliance. An extra function was added in 'stats.py'.