Sharing Of Appliances
Currently the load per appliance is calculated for each person and just added together. 
But this means that for example for the fridge the load is added multiple times. This does not represent the reality. 
Changes are made in residential.py and stats.py. In residential.py the load of appliances is calculated based on the type of appliance (shared or not, and continuously used or not). For this an extra function was added in stats.py.